Inventory/src/main/java/com/project/inventory/security/service/UserDetailsServiceImpl.java
package com.project.inventory.security.service;

import com.project.inventory.entity.Users;
import com.project.inventory.repository.UserRepo;
import jakarta.transaction.Transactional;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;
/**
 * userdetails service class.
*/

@Service
public class UserDetailsServiceImpl implements UserDetailsService {
  @Autowired
  UserRepo userRepository;
  
  public UserDetailsServiceImpl(UserRepo userRepository) {
    super();
    this.userRepository = userRepository;
  }

  @Override
  @Transactional
  public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
    Users user = userRepository.findByEmail(email)
        .orElseThrow(() -> new UsernameNotFoundException("User Not Found with email_id: " + email));

    return UserDetailsImpl.build(user);
  }
}
